More Updates.

+++ src/game/CreatureEventAI.cpp
@@ -303,7 +303,7 @@ bool CreatureEventAI::ProcessEvent(CreatureEventAIHolder& pHolder, Unit* pAction
             //Note: checked only aura for effect 0, if need check aura for effect 1/2 then
             // possible way: pack in event.buffed.amount 2 uint16 (ammount+effectIdx)
             Aura* aura = me->GetAura(event.buffed.spellId,0);
-            if (!aura || aura->GetStackAmount() < event.buffed.amount)
+            if (!aura || uint32(aura->GetStackAmount()) < event.buffed.amount)
                 return false;
 
             //Repeat Timers
@@ -319,7 +319,7 @@ bool CreatureEventAI::ProcessEvent(CreatureEventAIHolder& pHolder, Unit* pAction
             //Note: checked only aura for effect 0, if need check aura for effect 1/2 then
             // possible way: pack in event.buffed.amount 2 uint16 (ammount+effectIdx)
             Aura* aura = pActionInvoker->GetAura(event.buffed.spellId,0);
-            if (!aura || aura->GetStackAmount() < event.buffed.amount)
+            if (!aura || uint32(aura->GetStackAmount()) < event.buffed.amount)
                 return false;
 
             //Repeat Timers
@@ -331,7 +331,7 @@ bool CreatureEventAI::ProcessEvent(CreatureEventAIHolder& pHolder, Unit* pAction
             //Note: checked only aura for effect 0, if need check aura for effect 1/2 then
             // possible way: pack in event.buffed.amount 2 uint16 (ammount+effectIdx)
             Aura* aura = me->GetAura(event.buffed.spellId,0);
-            if (aura && aura->GetStackAmount() >= event.buffed.amount)
+            if (aura && uint32(aura->GetStackAmount()) >= event.buffed.amount)
                 return false;
 
             //Repeat Timers
@@ -344,7 +344,7 @@ bool CreatureEventAI::ProcessEvent(CreatureEventAIHolder& pHolder, Unit* pAction
                 return false;
 
             Aura* aura = me->getVictim()->GetAura(event.buffed.spellId,0);
-            if (aura && aura->GetStackAmount() >= event.buffed.amount)
+            if (aura && uint32(aura->GetStackAmount()) >= event.buffed.amount)
                 return false;
 
             //Repeat Timers
