+++ src/game/GameEventMgr.cpp
@@ -361,7 +361,7 @@ void GameEventMgr::LoadFromDB()
 
             int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
-            if (internal_event_id < 0 || internal_event_id >= mGameEventCreatureGuids.size())
+            if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventCreatureGuids.size())
             {
                 sLog.outErrorDb("game_event_creature game event id (%i) is out of range compared to max event id in game_event",event_id);
                 continue;
@@ -405,7 +405,7 @@ void GameEventMgr::LoadFromDB()
 
             int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
-            if (internal_event_id < 0 || internal_event_id >= mGameEventGameobjectGuids.size())
+            if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventGameobjectGuids.size())
             {
                 sLog.outErrorDb("game_event_gameobject game event id (%i) is out of range compared to max event id in game_event",event_id);
                 continue;
@@ -921,7 +921,7 @@ void GameEventMgr::LoadFromDB()
 
             int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
-            if (internal_event_id < 0 || internal_event_id >= mGameEventPoolIds.size())
+            if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventPoolIds.size())
             {
                 sLog.outErrorDb("`game_event_pool` game event id (%i) is out of range compared to max event id in `game_event`",event_id);
                 continue;
@@ -1136,7 +1136,7 @@ void GameEventMgr::GameEventSpawn(int16 event_id)
 {
     int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventCreatureGuids.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventCreatureGuids.size())
     {
         sLog.outError("GameEventMgr::GameEventSpawn attempt access to out of range mGameEventCreatureGuids element %i (size: %u)",internal_event_id,mGameEventCreatureGuids.size());
         return;
@@ -1164,7 +1164,7 @@ void GameEventMgr::GameEventSpawn(int16 event_id)
         }
     }
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventGameobjectGuids.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventGameobjectGuids.size())
     {
         sLog.outError("GameEventMgr::GameEventSpawn attempt access to out of range mGameEventGameobjectGuids element %i (size: %u)",internal_event_id,mGameEventGameobjectGuids.size());
         return;
@@ -1195,7 +1195,7 @@ void GameEventMgr::GameEventSpawn(int16 event_id)
         }
     }
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventPoolIds.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventPoolIds.size())
     {
         sLog.outError("GameEventMgr::GameEventSpawn attempt access to out of range mGameEventPoolIds element %i (size: %u)", internal_event_id, mGameEventPoolIds.size());
         return;
@@ -1213,7 +1213,7 @@ void GameEventMgr::GameEventUnspawn(int16 event_id)
 {
     int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventCreatureGuids.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventCreatureGuids.size())
     {
         sLog.outError("GameEventMgr::GameEventUnspawn attempt access to out of range mGameEventCreatureGuids element %i (size: %u)",internal_event_id,mGameEventCreatureGuids.size());
         return;
@@ -1234,7 +1234,7 @@ void GameEventMgr::GameEventUnspawn(int16 event_id)
         }
     }
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventGameobjectGuids.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventGameobjectGuids.size())
     {
         sLog.outError("GameEventMgr::GameEventUnspawn attempt access to out of range mGameEventGameobjectGuids element %i (size: %u)",internal_event_id,mGameEventGameobjectGuids.size());
         return;
@@ -1255,7 +1255,7 @@ void GameEventMgr::GameEventUnspawn(int16 event_id)
         }
     }
 
-    if (internal_event_id < 0 || internal_event_id >= mGameEventPoolIds.size())
+    if (internal_event_id < 0 || uint32(internal_event_id) >= mGameEventPoolIds.size())
     {
         sLog.outError("GameEventMgr::GameEventUnspawn attempt access to out of range mGameEventPoolIds element %i (size: %u)", internal_event_id, mGameEventPoolIds.size());
         return;
